+++ sites/funcampus/CHANGELOG.md
@@ -8,6 +8,19 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add middleware from richie.core to limit the browser cache TTL
+
+### Changed
+
+- Upgrade richie to 2.0.0-beta.15
+
+### Removed
+
+- Remove monkey patch that enabled cms page cache for non-staff users
+- Remove Django cache middlewares from the settings
+
 ## [0.4.1] - 2020-09-25
 
 ### Fixed